+++ CotEditor/Sources/Shortcut.swift
@@ -58,10 +58,6 @@ enum ModifierKey: CaseIterable {
     /// printable symbol
     var symbol: String {
         
-        if let symbol = self.sfSymbol {
-            return symbol
-        }
-        
         switch self {
             case .control: return "^"
             case .option:  return "⌥"
@@ -79,17 +75,6 @@ enum ModifierKey: CaseIterable {
             case .option:  return "~"
             case .shift:   return "$"
             case .command: return "@"
-        }
-    }
-    
-    
-    private var sfSymbol: String? {
-        
-        switch self {
-            case .control: return "􀆍"
-            case .option:  return "􀆕"
-            case .shift:   return "􀆝"
-            case .command: return "􀆔"
         }
     }
     
@@ -231,32 +216,9 @@ struct Shortcut: Hashable {
         
         guard let scalar = self.keyEquivalent.unicodeScalars.first else { return "" }
         
-        return Self.keyEquivalentSFSymbols[scalar]
-            ?? Self.keyEquivalentSymbols[scalar]
+        return Self.keyEquivalentSymbols[scalar]
             ?? self.keyEquivalent.uppercased()
     }
-    
-    
-    /// Key equivalent symbols that have SF Symbol alternatives.
-    private static let keyEquivalentSFSymbols: [Unicode.Scalar: String] = [
-        NSEvent.SpecialKey
-        .upArrow: "􀄤",
-        .downArrow: "􀄥",
-        .leftArrow: "􀄦",
-        .rightArrow: "􀄧",
-        .delete: "􁂒",
-        .backspace: "􁂈",
-        .home: "􀄿",
-        .end: "􀅀",
-        .pageUp: "􀄨",
-        .pageDown: "􀄩",
-        .clearLine: "􀆙",
-        .carriageReturn: "􀅇",
-        .enter: "􀆎",
-        .tab: "􁂎",
-        .backTab: "􁂊",
-        .escape: "􀆧",
-    ].mapKeys(\.unicodeScalar)
     
     
     /// table for key equivalent that have special symbols to display.