+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Change Log
 - [trivial] Improve Turkish localization (thanks to Emir SARI!).
 
 
+### Fixes
+
+- Fix an issue that the shortcut symbols in the Key Bindings preference pane did not display properly.
+
+
 
 4.1.0 (677)
 --------------------------